+++ shuup/admin/modules/users/views/password.py
@@ -130,7 +130,7 @@ class UserResetPasswordView(DetailView):
 
         r = RecoverPasswordForm()
         r.request = self.request
-        if r.process_user(user):
+        if r.process_user(user, self.request):
             messages.success(self.request, _(u"Password recovery email sent to %(email)s.") %
                              {"email": getattr(user, 'email', '')})
         else: